removed LM_EnumSymbolsEx and LM_FindSymbolsEx, because there's no need for a lm_process_t * parameter

+++ makebundle.sh
@@ -1,9 +1,16 @@
 #!/bin/sh
 
 mv libmem.a libmem_partial.a
-ln -s capstone-engine-prefix/src/capstone-engine-build/libcapstone.a ./libcapstone.a
-ln -s keystone-engine-prefix/src/keystone-engine-build/llvm/lib/libkeystone.a ./libkeystone.a
-ln -s lief-project-prefix/src/lief-project-build/libLIEF.a ./libLIEF.a
+
+if [ ! -f ./libcapstone.a ]; then
+	ln -s capstone-engine-prefix/src/capstone-engine-build/libcapstone.a ./libcapstone.a
+fi
+if [ ! -f ./libkeystone.a ]; then
+	ln -s keystone-engine-prefix/src/keystone-engine-build/llvm/lib/libkeystone.a ./libkeystone.a
+fi
+if [ ! -f ./libLIEF.a ]; then
+	ln -s lief-project-prefix/src/lief-project-build/libLIEF.a ./libLIEF.a
+fi
 ar -M << EOM
 	CREATE libmem.a
 	ADDLIB libmem_partial.a
